requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.36
alembic>=1.13.0
psycopg[binary]>=3.2.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
email-validator>=2.0.0
